move IssueId to its own module
